+++ libstorage/Storage.h
@@ -35,7 +35,7 @@ public:
 
     virtual ~Storage(){};
 
-    virtual Entries::Ptr select(h256 hash, int num, TableInfo::Ptr tableInfo,
+    virtual Entries::Ptr select(h256 hash, int64_t num, TableInfo::Ptr tableInfo,
         const std::string& key, Condition::Ptr condition = nullptr) = 0;
     virtual size_t commit(h256 hash, int64_t num, const std::vector<TableData::Ptr>& datas) = 0;
